+++ tests/syntax_test.rs
@@ -177,3 +177,46 @@ fn graft_basename() {
     assert!(value.is_some());
     assert_eq!("foo", value.unwrap());
 }
+
+#[test]
+fn escape_shell_variables() {
+    let value = syntax::escape_shell_variables("$");
+    assert_eq!(value, "$");
+
+    let value = syntax::escape_shell_variables("$ ");
+    assert_eq!(value, "$ ");
+
+    let value = syntax::escape_shell_variables("$$");
+    assert_eq!(value, "$$");
+
+    let value = syntax::escape_shell_variables("$_");
+    assert_eq!(value, "$$_");
+
+    let value = syntax::escape_shell_variables("$a");
+    assert_eq!(value, "$$a");
+
+    let value = syntax::escape_shell_variables("$_a");
+    assert_eq!(value, "$$_a");
+
+    let value = syntax::escape_shell_variables("$ echo");
+    assert_eq!(value, "$ echo");
+
+    let value = syntax::escape_shell_variables("embedded $$ value");
+    assert_eq!(value, "embedded $$ value");
+
+    let value = syntax::escape_shell_variables("$variable");
+    assert_eq!(value, "$$variable");
+
+    let value = syntax::escape_shell_variables("$$variable");
+    assert_eq!(value, "$$variable");
+
+    let value = syntax::escape_shell_variables("${braces}${ignored}");
+    assert_eq!(value, "${braces}${ignored}");
+
+    let value = syntax::escape_shell_variables("$a ${b} $c $");
+    assert_eq!(value, "$$a ${b} $$c $");
+
+    // Escaped ${braced} value
+    let value = syntax::escape_shell_variables("echo $${value[@]:0:1}");
+    assert_eq!(value, "echo $${value[@]:0:1}");
+}